Add tabbing between text fields
Allow the artist to tab between Find and then the Replace fields for
convenience.

Although there isn't a convention in Flame for tabbing between fields...
there is a standard convention of having focus on the main text input
field.  Willing to go outside of convention for this one to speed up
workflow.